Split resources into specifications & references
In the next commits, specifications will be made flexible to contain any names and be used in the handler declarations, while the exact resource references will be mapped from the cluster at runtime. One specification (in a handler) could match multiple actual resources (in the cluster).

But the first atomic step to this is the separation of the types, and the usage of them in the code accordingly. They both inherit from the same base class (also temporary) to share the same behaviour — changing it would be too much for now.

This split triggers the whole cascade of changes in how the registries are used: they are not structured as dicts with resources as keys anymore; instead, the registries are now flat, and the handler matching is done by the resource specification check — handler-by-handler.

**[BREAKING]** The old way of accessing the registries by the keys is supported for backward compatibility, but it returns the whole flat registry itself, which may contain handlers unrelated to the requested resource — proper methods should be used instead, which check for resource spec. If this is not done, handlers for unrelated resources can be exposed.